+++ cranelift/egraph/src/unionfind.rs
@@ -1,25 +1,27 @@
 //! Simple union-find data structure.
 
-use crate::Id;
-use cranelift_entity::EntityRef;
+use crate::{trace, Id};
+use cranelift_entity::SecondaryMap;
 
 /// A union-find data structure. The data structure can allocate
 /// `Id`s, indicating eclasses, and can merge eclasses together.
 #[derive(Clone, Debug)]
 pub struct UnionFind {
-    parent: Vec<Id>,
+    parent: SecondaryMap<Id, Id>,
 }
 
 impl UnionFind {
     /// Create a new `UnionFind`.
     pub fn new() -> Self {
-        UnionFind { parent: vec![] }
+        UnionFind {
+            parent: SecondaryMap::new(),
+        }
     }
 
     /// Create a new `UnionFind` with the given capacity.
     pub fn with_capacity(cap: usize) -> Self {
         UnionFind {
-            parent: Vec::with_capacity(cap),
+            parent: SecondaryMap::with_capacity(cap),
         }
     }
 
@@ -27,16 +29,13 @@ impl UnionFind {
     /// initially. All `Id`s must be added before being queried or
     /// unioned.
     pub fn add(&mut self, id: Id) {
-        if id.index() >= self.parent.len() {
-            self.parent.resize(id.index() + 1, Id::invalid());
-        }
-        self.parent[id.index()] = id;
+        self.parent[id] = id;
     }
 
     /// Find the canonical `Id` of a given `Id`.
     pub fn find(&self, mut node: Id) -> Id {
-        while node != self.parent[node.index()] {
-            node = self.parent[node.index()];
+        while node != self.parent[node] {
+            node = self.parent[node];
         }
         node
     }
@@ -48,12 +47,12 @@ impl UnionFind {
     pub fn find_and_update(&mut self, mut node: Id) -> Id {
         // "Path splitting" mutating find (Tarjan and Van Leeuwen).
         let orig = node;
-        while node != self.parent[node.index()] {
-            let next = self.parent[self.parent[node.index()].index()];
-            self.parent[node.index()] = next;
+        while node != self.parent[node] {
+            let next = self.parent[self.parent[node]];
+            self.parent[node] = next;
             node = next;
         }
-        log::trace!("find_and_update: {} -> {}", orig, node);
+        trace!("find_and_update: {} -> {}", orig, node);
         node
     }
 
@@ -64,8 +63,8 @@ impl UnionFind {
         let (a, b) = (std::cmp::min(a, b), std::cmp::max(a, b));
         if a != b {
             // Always canonicalize toward lower IDs.
-            self.parent[b.index()] = a;
-            log::trace!("union: {}, {}", a, b);
+            self.parent[b] = a;
+            trace!("union: {}, {}", a, b);
         }
     }
 }
