+++ cranelift/entity/src/list.rs
@@ -137,24 +137,6 @@ impl<T: EntityRef + ReservedValue> ListPool<T> {
         }
     }
 
-    /// Create a new list pool with the given capacity for data pre-allocated.
-    pub fn with_capacity(len: usize) -> Self {
-        Self {
-            data: Vec::with_capacity(len),
-            free: Vec::new(),
-        }
-    }
-
-    /// Get the capacity of this pool. This will be somewhat higher
-    /// than the total length of lists that can be stored without
-    /// reallocating, because of internal metadata overheads. It is
-    /// mostly useful to allow another pool to be allocated that is
-    /// likely to hold data transferred from this one without the need
-    /// to grow.
-    pub fn capacity(&self) -> usize {
-        self.data.capacity()
-    }
-
     /// Clear the pool, forgetting about all lists that use it.
     ///
     /// This invalidates any existing entity lists that used this pool to allocate memory.
@@ -473,53 +455,6 @@ impl<T: EntityRef + ReservedValue> EntityList<T> {
             for x in iterator {
                 self.push(x, pool);
             }
-        }
-    }
-
-    /// Appends another list from the same ListPool onto this list.
-    pub fn append_list(&mut self, other: &EntityList<T>, pool: &mut ListPool<T>) {
-        let offset = self.len(pool);
-        let new_count = other.len(pool);
-        debug_assert!(new_count <= usize::MAX - offset);
-        self.grow(new_count, pool);
-        for i in 0..new_count {
-            let elt = other.get(i, pool).unwrap();
-            *self.get_mut(offset + i, pool).unwrap() = elt;
-        }
-    }
-
-    /// Sorts the list.
-    pub fn sort(&mut self, pool: &mut ListPool<T>)
-    where
-        T: Ord,
-    {
-        self.as_mut_slice(pool).sort_unstable();
-    }
-
-    /// Removes duplicates from the list. Assumes the list has already
-    /// been sorted.
-    pub fn remove_dups(&mut self, pool: &mut ListPool<T>) {
-        let mut last = None;
-        let mut out = 0;
-        let data = self.as_mut_slice(pool);
-        for i in 0..data.len() {
-            if last == Some(data[i]) {
-                continue;
-            }
-            last = Some(data[i]);
-
-            if out < i {
-                data[out] = data[i];
-            }
-            out += 1;
-        }
-
-        // If we had dups, shrink list length.
-        if out < data.len() {
-            debug_assert!(data.len() > 0); // Empty list will not reach here.
-            debug_assert!(out > 0); // Non-empty list cannot become empty.
-            debug_assert!(self.index > 0);
-            pool.data[(self.index - 1) as usize] = T::new(out);
         }
     }
